#17 channelの名前だけにする
