+++ brcore/enum/__init__.py
@@ -1,0 +1,5 @@
+from brcore.enum.exception_texts import ExceptionTexts
+from brcore.enum.channels import MisskeyChannelNames
+
+
+__all__ = ["ExceptionTexts", "MisskeyChannelNames"]
